Minimize duplicate state
I could not remove the selectedImageIndex completly, because if i have a function in the image surce : src={`${src}/${getSelectedImageIndex()}`} the function is called once and sets its value. This works for the image in the readingmode, but not for the image in the editing mode. In the editing mode the current selected Image can change by calling the saveSelectedImage from the UI. The changes are then correctly stored at the backend but because the src dont call the function getSelectedImageIndex() again it is not shown in the UI. Therefore I still have to use a variable that gets updated when saveSelectedImage(version) is called.